+++ node_helper.js
@@ -1,7 +1,5 @@
 var NodeHelper = require('node_helper');
 const axios = require('axios');
-const es = require('./es.js');
-const pt = require('./pt.js');
 const wotd = require('./de.js');
 
 module.exports = NodeHelper.create ({
@@ -9,28 +7,12 @@ module.exports = NodeHelper.create ({
         console.log("Starting node_helper for: " + this.name);
     },
 
-    getESData: function() {
-        var self = this;
-
-        es.getesData(function(translationData) {
-            self.sendSocketNotification("WOTD_ESP_DATA", translationData)
-        });
-    },
-
-    getPTData: function() {
-        var self = this;
-        
-        pt.getptData(function(translationData) {
-            self.sendSocketNotification("WOTD_PT_DATA", translationData)
-        });        
-    },
-
     getData: function(url) {
         var self = this;
         
-        wotd.getData(function(translationData, url) {
+        wotd.getData(function(translationData) {
             self.sendSocketNotification("WOTD_DATA", translationData)
-        });        
+        }, url);        
     },
 
     socketNotificationReceived: function(notification, payload) {
