+++ wotd.js
@@ -2,17 +2,17 @@ var axios = require('axios');
 var cheerio = require('cheerio');
 
 module.exports = {
-    url: "https://www.portuguesepod101.com/portuguese-phrases/",
+    // url: "https://www.germanpod101.com/german-phrases/",
 
-    getPorData: function($) {
-        let element = {}, portugueseWords = [];
+    getDEData: function($) {
+        let element = {}, germanWords = [];
         $(".r101-wotd-widget__word").each((index, p) => {
             element.id = index;
             element.word = p.firstChild.data;
-            portugueseWords.push(element.word);
+            germanWords.push(element.word);
         })
 
-        return portugueseWords;
+        return germanWords;
     },
 
     getEngData: function($) {
@@ -26,21 +26,22 @@ module.exports = {
         return englishWords;
     },
 
-    getptData: function(callback) {
-        axios.get(this.url).then(({ data }) => {
+    getData: function(callback, obj) {
+        url = obj
+        axios.get(url).then(({ data }) => {
             const $ = cheerio.load(data, null, true);
 
             let translationData = [];
 
-            const portuguese = this.getPorData($);
+            const german = this.getDEData($);
 
             const english = this.getEngData($);
 
             translationData.push({
-                "word": portuguese[0],
+                "word": german[0],
                 "translation": english[0],
                 "examples": {
-                    "wordex": portuguese[1],
+                    "wordex": german[1],
                     "wordextr": english[1],
                 }
             })
